+++ packages/perspective/src/js/config/constants.js
@@ -79,9 +79,3 @@ export const TYPE_FILTERS = {
     datetime: DATETIME_FILTERS,
     date: DATETIME_FILTERS
 };
-
-export const TYPED_ARRAY_SENTINEL_VALUE_INT8 = 255;
-export const TYPED_ARRAY_SENTINEL_VALUE_INT16 = 32767;
-export const TYPED_ARRAY_SENTINEL_VALUE_INT32 = 2147483647;
-export const TYPED_ARRAY_SENTINEL_VALUE_FLOAT32 = 3.40282e38;
-export const TYPED_ARRAY_SENTINEL_VALUE_FLOAT64 = Number.MAX_VALUE;
